JDK-8251513: Code around Parse::do_lookupswitch/do_tableswitch should be cleaned up

+++ src/hotspot/share/ci/ciStreams.hpp
@@ -203,8 +203,8 @@ public:
   }
 
   // For a lookup or switch table, return target destination
-  int get_int_table( int index ) const {
-    return Bytes::get_Java_u4((address)&_table_base[index]); }
+  jint get_int_table( int index ) const {
+    return (jint)Bytes::get_Java_u4((address)&_table_base[index]); }
 
   int get_dest_table( int index ) const {
     return cur_bci() + get_int_table(index); }
